reorganise stories
