+++ src/components/Customs/Selects/SyAutocomplete/SyAutocomplete.stories.ts
@@ -2,7 +2,7 @@ import type { Meta, StoryObj } from '@storybook/vue3'
 import SyAutocomplete from './SyAutocomplete.vue'
 import SyForm from '../../SyForm/SyForm.vue'
 import AccessibilityDocs from './accessibilite/Accessibility.mdx'
-import { ref, onMounted } from 'vue'
+import { ref } from 'vue'
 import { fn } from '@storybook/test'
 import { VBtn } from 'vuetify/components'
 import { getValidationDocumentation } from '@/composables/unifyValidation/documentationValidationProps'
@@ -423,293 +423,6 @@ const items = [
 			`,
 		}
 	},
-}
-
-export const WithError: Story = {
-	parameters: {
-		docs: {
-			description: {
-				story: '« Option 1 » est présélectionnée et déclenche une erreur bloquante au chargement.',
-			},
-		},
-		sourceCode: [
-			{
-				name: 'Template',
-				code: `
-<template>
-  <SyAutocomplete
-    ref="autocompleteRef"
-    v-model="value"
-    :items="items"
-    label="Option"
-    :customRules="[
-      {
-        type: 'custom',
-        options: {
-          validate: (v) => v !== '1',
-          message: 'Option 1 n\\'est pas autorisée, choisissez Option 2 ou 3.'
-        }
-      }
-    ]"
-  />
-</template>`,
-			},
-			{
-				name: 'Script',
-				code: `
-<script setup lang="ts">
-import { onMounted, ref } from 'vue'
-import { SyAutocomplete } from '@cnamts/synapse'
-
-const value = ref('1')
-const items = [
-  { text: 'Option 1', value: '1' },
-  { text: 'Option 2', value: '2' },
-  { text: 'Option 3', value: '3' },
-]
-
-const autocompleteRef = ref(null)
-
-onMounted(() => {
-  autocompleteRef.value?.validateOnSubmit()
-})
-</script>`,
-			},
-		],
-	},
-	args: {
-		'items': [
-			{ text: 'Option 1', value: '1' },
-			{ text: 'Option 2', value: '2' },
-			{ text: 'Option 3', value: '3' },
-		],
-		'label': 'Option',
-		'filter': false,
-		'onUpdate:modelValue': fn(),
-	},
-	render: args => ({
-		components: { SyAutocomplete },
-		setup() {
-			const value = ref('1')
-			const autocompleteRef = ref<{ validateOnSubmit: () => Promise<boolean> } | null>(null)
-
-			onMounted(() => {
-				autocompleteRef.value?.validateOnSubmit()
-			})
-
-			return { args, value, autocompleteRef }
-		},
-		template: `
-			<div class="pa-4">
-				<SyAutocomplete
-					ref="autocompleteRef"
-					v-model="value"
-					v-bind="args"
-					:customRules="[
-						{
-							type: 'custom',
-							options: {
-								validate: (v) => v !== '1',
-								message: 'Option 1 n\\'est pas autorisée, choisissez Option 2 ou 3.'
-							}
-						}
-					]"
-				/>
-			</div>
-		`,
-	}),
-}
-
-export const WithWarning: Story = {
-	parameters: {
-		docs: {
-			description: {
-				story: '« Option 1 » est présélectionnée et déclenche un avertissement (non bloquant) au chargement.',
-			},
-		},
-		sourceCode: [
-			{
-				name: 'Template',
-				code: `
-<template>
-  <SyAutocomplete
-    ref="autocompleteRef"
-    v-model="value"
-    :items="items"
-    label="Option"
-    :customWarningRules="[
-      {
-        type: 'custom',
-        options: {
-          validate: (v) => v !== '1',
-          warningMessage: 'Option 1 est dépréciée, préférez Option 2 ou 3.'
-        }
-      }
-    ]"
-  />
-</template>`,
-			},
-			{
-				name: 'Script',
-				code: `
-<script setup lang="ts">
-import { onMounted, ref } from 'vue'
-import { SyAutocomplete } from '@cnamts/synapse'
-
-const value = ref('1')
-const items = [
-  { text: 'Option 1', value: '1' },
-  { text: 'Option 2', value: '2' },
-  { text: 'Option 3', value: '3' },
-]
-
-const autocompleteRef = ref(null)
-
-onMounted(() => {
-  autocompleteRef.value?.validateOnSubmit()
-})
-</script>`,
-			},
-		],
-	},
-	args: {
-		'items': [
-			{ text: 'Option 1', value: '1' },
-			{ text: 'Option 2', value: '2' },
-			{ text: 'Option 3', value: '3' },
-		],
-		'label': 'Option',
-		'filter': false,
-		'onUpdate:modelValue': fn(),
-	},
-	render: args => ({
-		components: { SyAutocomplete },
-		setup() {
-			const value = ref('1')
-			const autocompleteRef = ref<{ validateOnSubmit: () => Promise<boolean> } | null>(null)
-
-			onMounted(() => {
-				autocompleteRef.value?.validateOnSubmit()
-			})
-
-			return { args, value, autocompleteRef }
-		},
-		template: `
-			<div class="pa-4">
-				<SyAutocomplete
-					ref="autocompleteRef"
-					v-model="value"
-					v-bind="args"
-					:customWarningRules="[
-						{
-							type: 'custom',
-							options: {
-								validate: (v) => v !== '1',
-								warningMessage: 'Option 1 est dépréciée, préférez Option 2 ou 3.'
-							}
-						}
-					]"
-				/>
-			</div>
-		`,
-	}),
-}
-
-export const WithSuccess: Story = {
-	parameters: {
-		docs: {
-			description: {
-				story: 'Une option est présélectionnée et déclenche la confirmation de succès au chargement.',
-			},
-		},
-		sourceCode: [
-			{
-				name: 'Template',
-				code: `
-<template>
-  <SyAutocomplete
-    ref="autocompleteRef"
-    v-model="value"
-    :items="items"
-    label="Option"
-    show-success-messages
-    :customSuccessRules="[
-      {
-        type: 'custom',
-        options: {
-          validate: (v) => v !== null && v !== undefined,
-          successMessage: 'Option sélectionnée avec succès.'
-        }
-      }
-    ]"
-  />
-</template>`,
-			},
-			{
-				name: 'Script',
-				code: `
-<script setup lang="ts">
-import { onMounted, ref } from 'vue'
-import { SyAutocomplete } from '@cnamts/synapse'
-
-const value = ref('1')
-const items = [
-  { text: 'Option 1', value: '1' },
-  { text: 'Option 2', value: '2' },
-  { text: 'Option 3', value: '3' },
-]
-
-const autocompleteRef = ref(null)
-
-onMounted(() => {
-  autocompleteRef.value?.validateOnSubmit()
-})
-</script>`,
-			},
-		],
-	},
-	args: {
-		'items': [
-			{ text: 'Option 1', value: '1' },
-			{ text: 'Option 2', value: '2' },
-			{ text: 'Option 3', value: '3' },
-		],
-		'label': 'Option',
-		'filter': false,
-		'showSuccessMessages': true,
-		'onUpdate:modelValue': fn(),
-	},
-	render: args => ({
-		components: { SyAutocomplete },
-		setup() {
-			const value = ref('1')
-			const autocompleteRef = ref<{ validateOnSubmit: () => Promise<boolean> } | null>(null)
-
-			onMounted(() => {
-				autocompleteRef.value?.validateOnSubmit()
-			})
-
-			return { args, value, autocompleteRef }
-		},
-		template: `
-			<div class="pa-4">
-				<SyAutocomplete
-					ref="autocompleteRef"
-					v-model="value"
-					v-bind="args"
-					:customSuccessRules="[
-						{
-							type: 'custom',
-							options: {
-								validate: (v) => v !== null && v !== undefined,
-								successMessage: 'Option sélectionnée avec succès.'
-							}
-						}
-					]"
-				/>
-			</div>
-		`,
-	}),
 }
 
 export const ReturnObjectWithCustomKeys: Story = {
@@ -1228,88 +941,6 @@ const value2 = ref('1')
 	},
 }
 
-export const DisableErrorHandling: Story = {
-	parameters: {
-		sourceCode: [
-			{
-				name: 'Template',
-				code: `
-<template>
-	<div class="d-flex flex-column gap-4">
-		<SyAutocomplete
-			v-model="value1"
-			:items="items"
-			label="Avec validation interne (défaut)"
-			required
-		/>
-
-		<SyAutocomplete
-			v-model="value2"
-			:items="items"
-			label="Validation interne désactivée"
-			required
-			disable-error-handling
-		/>
-	</div>
-</template>
-				`,
-			},
-			{
-				name: 'Script',
-				code: `
-<script setup lang="ts">
-import { ref } from 'vue'
-import { SyAutocomplete } from '@cnamts/synapse'
-
-const items = [
-	{ text: 'Option 1', value: '1' },
-	{ text: 'Option 2', value: '2' },
-	{ text: 'Option 3', value: '3' },
-]
-
-const value1 = ref(null)
-const value2 = ref(null)
-</script>
-				`,
-			},
-		],
-	},
-	args: {
-		items: [
-			{ text: 'Option 1', value: '1' },
-			{ text: 'Option 2', value: '2' },
-			{ text: 'Option 3', value: '3' },
-		],
-	},
-	render: (args) => {
-		return {
-			components: { SyAutocomplete },
-			setup() {
-				const value1 = ref(null)
-				const value2 = ref(null)
-				return { args, value1, value2 }
-			},
-			template: `
-				<div class="pa-4 d-flex flex-column" style="gap: 16px;">
-					<SyAutocomplete
-						v-model="value1"
-						v-bind="args"
-						label="Avec validation interne (défaut)"
-						required
-					/>
-					<SyAutocomplete
-						v-model="value2"
-						v-bind="args"
-						label="Validation interne désactivée"
-						required
-						disable-error-handling
-					/>
-				</div>
-			`,
-		}
-	},
-}
-
 export const ReadonlyField: Story = {
 	parameters: {
 
@@ -1498,201 +1129,4 @@ const items = [
 			`,
 		}
 	},
-}
-
-export const FormValidation: Story = {
-	parameters: {
-		docs: {
-			description: {
-				story: 'Le champ requis ne montre l\'erreur qu\'après interaction (blur/submit), pas au mount.',
-			},
-		},
-		sourceCode: [
-			{
-				name: 'Template',
-				code: `
-<template>
-  <SyForm @submit="onSubmit">
-    <SyAutocomplete
-      v-model="value"
-      :items="items"
-      label="Recherche obligatoire"
-      required
-      display-asterisk
-      class="mb-4"
-    />
-    <VBtn type="submit" color="primary">Soumettre</VBtn>
-  </SyForm>
-</template>
-        `,
-			},
-			{
-				name: 'Script',
-				code: `
-<script setup lang="ts">
-import { ref } from 'vue'
-import { SyAutocomplete, SyForm } from '@cnamts/synapse'
-import { VBtn } from 'vuetify/components'
-
-const value = ref('')
-const items = [
-  { text: 'Option 1', value: '1' },
-  { text: 'Option 2', value: '2' },
-  { text: 'Option 3', value: '3' }
-]
-
-const onSubmit = (event) => {
-  if (event.isValid) {
-    alert('Formulaire valide : ' + JSON.stringify(value.value))
-  } else {
-    alert('Formulaire invalide : veuillez choisir une option.')
-  }
-}
-</script>
-        `,
-			},
-		],
-	},
-	args: {
-		items: [
-			{ text: 'Option 1', value: '1' },
-			{ text: 'Option 2', value: '2' },
-			{ text: 'Option 3', value: '3' },
-		],
-		label: 'Recherche obligatoire',
-		required: true,
-		displayAsterisk: true,
-	},
-	render: (args) => {
-		return {
-			components: { SyAutocomplete, SyForm, VBtn },
-			setup() {
-				const value = ref('')
-
-				const onSubmit = (event: { isValid: boolean }) => {
-					if (event.isValid) {
-						alert(`Formulaire valide : ${JSON.stringify(value.value)}`)
-					}
-					else {
-						alert('Formulaire invalide : veuillez choisir une option.')
-					}
-				}
-
-				return { args, value, onSubmit }
-			},
-			template: `
-				<div class="pa-4">
-					<SyForm @submit="onSubmit">
-						<SyAutocomplete
-							v-model="value"
-							v-bind="args"
-							class="mb-4"
-						/>
-						<VBtn type="submit" color="primary">Soumettre</VBtn>
-					</SyForm>
-				</div>
-			`,
-		}
-	},
-}
-
-export const VuetifyValidation: Story = {
-	parameters: {
-		docs: {
-			description: {
-				story: 'Exemple d\'utilisation de la validation native Vuetify via la prop `useVuetifyValidation`. Les règles sont définies au format Vuetify : des fonctions retournant `true` ou un message d\'erreur. Soumettez le formulaire pour déclencher la validation.',
-			},
-		},
-		sourceCode: [
-			{
-				name: 'Template',
-				code: `
-<template>
-  <SyForm @submit="onSubmit">
-    <SyAutocomplete
-      v-model="value"
-      :items="items"
-      label="Option"
-      use-vuetify-validation
-      :show-success-messages="false"
-      :rules="[v => !!v || 'Ce champ est requis']"
-    />
-    <VBtn type="submit" color="primary" class="mt-4">
-      Soumettre
-    </VBtn>
-  </SyForm>
-</template>`,
-			},
-			{
-				name: 'Script',
-				code: `
-<script setup lang="ts">
-import { ref } from 'vue'
-import { SyAutocomplete, SyForm } from '@cnamts/synapse'
-import { VBtn } from 'vuetify/components'
-
-const value = ref(null)
-const items = [
-  { text: 'Option 1', value: '1' },
-  { text: 'Option 2', value: '2' },
-  { text: 'Option 3', value: '3' },
-]
-
-function onSubmit(event: { isValid: boolean }) {
-  if (event.isValid) {
-    alert('Formulaire valide : ' + value.value)
-  } else {
-    alert('Formulaire invalide : veuillez choisir une option.')
-  }
-}
-</script>`,
-			},
-		],
-	},
-	args: {
-		'items': [
-			{ text: 'Option 1', value: '1' },
-			{ text: 'Option 2', value: '2' },
-			{ text: 'Option 3', value: '3' },
-		],
-		'label': 'Option',
-		'useVuetifyValidation': true,
-		'showSuccessMessages': false,
-		'onUpdate:modelValue': fn(),
-	},
-	render: args => ({
-		components: { SyAutocomplete, SyForm, VBtn },
-		setup() {
-			const value = ref(null)
-
-			function onSubmit(event: { isValid: boolean }) {
-				if (event.isValid) {
-					alert(`Formulaire valide : ${value.value}`)
-				}
-				else {
-					alert('Formulaire invalide : veuillez choisir une option.')
-				}
-			}
-
-			return { args, value, onSubmit }
-		},
-		template: `
-			<div class="pa-4">
-				<SyForm @submit="onSubmit">
-					<SyAutocomplete
-						v-model="value"
-						v-bind="args"
-						:rules="[v => !!v || 'Ce champ est requis']"
-					/>
-					<VBtn
-						type="submit"
-						color="primary"
-						class="mt-4"
-					>
-						Soumettre
-					</VBtn>
-				</SyForm>
-			</div>
-		`,
-	}),
 }
